requirements.txt
billboard.py==6.1.0
boto3==1.13.3
numpy==1.18.1
numpydoc==0.9.2
pandas==1.0.3
spotipy==2.10.0
SQLAlchemy==1.3.15
